// types.ts
export type Movie = {
    _id: string;
    Title: string;
    Year: string;
    Rated: string;
    Released: string;
    Runtime: string;
    Genre: string;
    Director: string;
    $vectorize: string;
    $vector: Array<number>;
    Writer: string;
    Actors: string;
    Language: string;
    Country: string;
    Awards: string;
    Poster: string;
    Metascore: string;
    imdbRating: string;
    imdbVotes: string;
    imdbID: string;
    Type: string;
    DVD: string;
    BoxOffice: string;
    Production: string;
    Website: string;
    Response: string;
  };
  
  export type SimilarMovie = {
    $similarity: number;
  } & Movie;